Refactoring and creation of open_ai folder
